requirements.txt
absl-py==0.9.0
alabaster==0.7.12
altgraph==0.17
anaconda-client==1.7.2
anaconda-navigator==1.9.12
appdirs==1.4.4
applaunchservices==0.2.1
appnope==0.1.0
argh==0.26.2
asgiref==3.2.10
asn1crypto==1.3.0
astor==0.7.1
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
auto-py-to-exe==2.7.11
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
clyent==1.2.2
conda==4.9.2
conda-build==3.17.8
conda-package-handling==1.6.0
conda-verify==3.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-allauth==0.44.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
docutils==0.16
drf-yasg==1.20.0
Eel==0.12.4
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
future==0.18.2
gast==0.3.3
gevent==20.9.0
gevent-websocket==0.10.1
glob2==0.7
graphviz==0.14.1
greenlet==1.0a1
grpcio==1.23.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
inflection==0.5.1
iniconfig==1.1.1
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itypes==1.2.0
jedi==0.15.2
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.3
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
lief==0.9.0
macholib==1.14
Mako==1.1.0
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1589366472132/work
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
notebook==6.0.3
numpy==1.18.2
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.5.2
pathtools==0.1.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pip-review==1.1.0
pipenv==2020.11.15
pipupgrade==1.8.0
pkginfo==1.5.0.1
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.9.2
psutil==5.7.0
psycopg2==2.7.1
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.10.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.6.1
pygpu==0.7.6
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
PyJWT==2.0.0
pylint==2.4.4
Pympler==0.7
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
pytest==6.2.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.7.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.16.0
Rtree==0.9.3
ruamel-yaml==0.15.87
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1610147049074/work
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn @ file:///tmp/build/80754af9/seaborn_1600553570093/work
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
soupsieve==2.0
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.1
spyder-kernels==1.9.0
sqlparse==0.3.0
statsmodels==0.11.0
swagger-spec-validator==2.7.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
Theano==1.0.4
toml==0.10.2
tornado==6.0.4
tqdm==4.44.1
traitlets==4.3.3
typed-ast==1.4.0
ujson==1.35
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.2.2
virtualenv-clone==0.5.4
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
whichcraft==0.6.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.0
xmltodict==0.12.0
yapf==0.28.0
zipp==2.2.0
zope.event==4.5.0
zope.interface==5.2.0
